Electron producer and service of transform session
